More zombie changes :pear:
Changed the zombie's sprite to actually be a zombie sprite.

changed the draw code to use the sprite index.

when zombies choose what "type" they are, they now choose up to the last frame of the current sprites animation. this will prevent us from having to go back and change it ever, even if we change it later on. :thumbsup: